Move constants and "preamble" into different files
This is in preperation for creating the grid that the displays stand on.
I will need the constants that are shared across the grid and display in
a seperate file, so that I can `include` it and can safely `use` the
display file to get the `grid_foot`.

I also added a preamble which sets `$fn` and then `include`s the
constants which is meant to be `include`d in every file of this project.
(I feel that setting `$fn` in the constants file is a bit weird ^^)

I have also added `four_corners` to the preamble as I think of it as a
utility function.